move librs.py from unit-tests/dds/ to unit-tests/py/rspy
