+++ unit-tests/dds/test-librs-extrinsics.py
@@ -7,7 +7,7 @@
 from rspy import log, test
 log.nested = 'C  '
 
-import librs as rs
+from rspy import librs as rs
 if log.is_debug_on():
     rs.log_to_console( rs.log_severity.debug )
 from time import sleep
