+++ unit-tests/dds/test-librs-options.py
@@ -47,7 +47,7 @@ with test.remote.fork( nested_indent=None ) as remote:
     ###############################################################################################################
     # The client is LibRS
     #
-    import librs as rs
+    from rspy import librs as rs
     if log.is_debug_on():
         rs.log_to_console( rs.log_severity.debug )
